+++ apps/passport-server/src/services/generic-issuance/capabilities/types.ts
@@ -1,0 +1,8 @@
+/**
+ * {@link Pipeline}s can have external-facing APIs. Anything that an external
+ * user can do with a Pipeline needs to be represented as a {@link PipelineCapability}.
+ */
+export enum PipelineCapability {
+  FeedIssuanceCapability = "FeedIssuanceCapability",
+  CheckinCapability = "CheckinCapability"
+}
