+++ apps/passport-server/src/services/generic-issuance/pipelines/PretixPipeline.ts
@@ -1,6 +1,11 @@
 import { PipelineAtomDB } from "../../../database/queries/pipelineAtomDB";
-import { PipelineType, PretixPipelineDefinition } from "../types";
-import { BasePipeline, Pipeline } from "./types";
+import {
+  BasePipeline,
+  BasePipelineDefinition,
+  Pipeline,
+  PipelineDefinition,
+  PipelineType
+} from "./types";
 
 /**
  * TODO: implement this. (Probably Rob).
@@ -26,4 +31,42 @@ export class PretixPipeline implements BasePipeline {
   public static is(p: Pipeline): p is PretixPipeline {
     return p.type === PipelineType.Pretix;
   }
+}
+
+/**
+ * Similar to {@link LemonadePipelineDefinition} but for Pretix-based Pipelines.
+ */
+export interface PretixPipelineDefinition extends BasePipelineDefinition {
+  type: PipelineType.Pretix;
+  options: PretixPipelineOptions;
+}
+
+export function isPretixPipelineDefinition(
+  d: PipelineDefinition
+): d is PretixPipelineDefinition {
+  return d.type === PipelineType.Pretix;
+}
+
+/**
+ * TODO: this needs to take into account the actual {@link PretixPipeline}, which
+ * has not been implemented yet.
+ */
+export interface PretixPipelineOptions {
+  pretixAPIKey: string;
+  pretixOrgUrl: string;
+  events: PretixEventConfig[];
+}
+
+/**
+ * This object represents a configuration from which the server can instantiate
+ * a functioning {@link PretixPipeline}. It's entirely specified by the user.
+ *
+ * TODO:
+ * - how do these map to product and event ids?
+ */
+export interface PretixEventConfig {
+  id: string;
+  name: string;
+  productIds: string[];
+  superUserProductIds: string[];
 }
